etl fix 8

+++ acs/acs_ygf_place_of_birth_for_foreing_born/helper.py
@@ -1,6 +1,6 @@
 def add_cols_2019(df):
     for col in df.columns:
-        df[col] = df[col].astype(int) 
+        df[col] = df[col].astype(float) 
 
     if "B05006_099E" in df.columns:
         df["B05006_099E"] = df["B05006_097E"] + df["B05006_098E"] + df["B05006_099E"]
